+++ app/controllers/application_controller.rb
@@ -31,15 +31,6 @@ class ApplicationController < ActionController::Base
     current_user && current_user.admin?
   end
 
-
-  def current_user
-    @current_user ||= User.find(session[:user_id]) if session[:user_id]
-  end
-
-  def current_admin?
-    current_user && current_user.admin?
-  end
-
   def price_in_dollars(price)
     "$#{sprintf('%.2f', price/100)}"
   end
@@ -48,16 +39,9 @@ class ApplicationController < ActionController::Base
     User.states
   end
 
-
-
-
-
   def logged_in?
     !current_user.nil?
   end
-
-
-
 
     private
