+++ app/controllers/categories_controller.rb
@@ -6,7 +6,7 @@ class CategoriesController < ApplicationController
 
   def show
     @category = Category.find_by(slug: params[:category])
-    render_404 and return if @category.nil?
+    render file: "/public/404" if @category.nil?
     @items = @category.items.paginate(:page => params[:page], :per_page => 15)
     render "items/index"
   end
